rename package: config (#389)
