--- exo-go/src/app_config_helpers/index_suite_test.go
@@ -1,13 +1,0 @@
-package appConfigHelpers_test
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-)
-
-func TestAppConfigHelpers(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "AppConfigHelpers Suite")
-}
