--- exo-go/src/app_dependency_helpers/index_suite_test.go
@@ -1,13 +1,0 @@
-package appDependencyHelpers_test
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-)
-
-func TestTypes(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "AppDependencyHelpers Suite")
-}
