+++ exo-go/src/config/generic_dependency.go
@@ -1,4 +1,4 @@
-package appDependencyHelpers
+package config
 
 import (
 	"fmt"
