+++ exo-go/src/config/nats_dependency.go
@@ -1,4 +1,4 @@
-package appDependencyHelpers
+package config
 
 import (
 	"fmt"
