+++ exo-go/src/config/service_config.go
@@ -1,4 +1,4 @@
-package serviceConfigHelpers
+package config
 
 import (
 	"fmt"
@@ -43,23 +43,11 @@ func getInternalServiceConfig(appDir, serviceDirName string) (types.ServiceConfi
 	return serviceConfig, nil
 }
 
-// GetServiceData returns the configurations data for the given services
-func GetServiceData(services types.Services) map[string]types.ServiceData {
-	result := make(map[string]types.ServiceData)
-	for serviceName, data := range services.Private {
-		result[serviceName] = data
-	}
-	for serviceName, data := range services.Public {
-		result[serviceName] = data
-	}
-	return result
-}
-
 // GetServiceConfigs reads the service.yml of all services and returns
 // the serviceConfig objects and an error (if any)
 func GetServiceConfigs(appDir string, appConfig types.AppConfig) (map[string]types.ServiceConfig, error) {
 	result := map[string]types.ServiceConfig{}
-	for service, serviceData := range GetServiceData(appConfig.Services) {
+	for service, serviceData := range appConfig.GetServiceData() {
 		if len(serviceData.Location) > 0 {
 			serviceConfig, err := getInternalServiceConfig(appDir, serviceData.Location)
 			if err != nil {
