+++ exo-go/src/config/suite_test.go
@@ -1,0 +1,24 @@
+package config_test
+
+import (
+	"testing"
+
+	"github.com/Originate/exosphere/exo-go/src/util"
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestAppConfigHelpers(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Config Suite")
+}
+
+var homeDir string
+
+var _ = BeforeSuite(func() {
+	var err error
+	homeDir, err = util.GetHomeDirectory()
+	if err != nil {
+		panic(err)
+	}
+})
