--- exo-go/src/service_config_helpers/index_suite_test.go
@@ -1,13 +1,0 @@
-package serviceConfigHelpers_test
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-)
-
-func TestServiceConfigHelpers(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "ServiceConfigHelpers Suite")
-}
